Initial impl of KNN (K-Nearest Neighbor).
Note:  the skeleton package is sample code given to the class.  At this commit, the KNearestNeighbor_Skeleton.java is being used as the entry point and launcher.  This will change, and eventually the user will specify the k-value at run time on the command line as an argument to java.